requirements.txt
boto3==1.20.53
botocore==1.23.53
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
jmespath==0.10.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
s3transfer==0.5.1
six==1.16.0
urllib3==1.26.8
